Split internal-links related code in a separate file
No functional or code changes. Only moving the code and tests to
a separate file then importing them from the new location.
